[Fix][修复部分添加主线的BUG]
